Adding unit tests for command and queries directly to ensure aid in adding validation rules

+++ babysitter/src/BabySitter.Core/BabySitters/Models/SitterModel.cs
@@ -5,8 +5,8 @@
         public int Id { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
-        public int HourlyRate { get; set; }
-        public int HourlyRateAfterMidnight { get; set; }
-        public int HourlyRateBetweenBedtimeAndMidnight { get; set; }
+        public long HourlyRate { get; set; }
+        public long HourlyRateAfterMidnight { get; set; }
+        public long HourlyRateBetweenBedtimeAndMidnight { get; set; }
     }
 }
