+++ babysitter/src/BabySitter.Core/BabySitters/Shifts/Models/ShiftModel.cs
@@ -9,8 +9,8 @@ namespace BabySitter.Core.BabySitters.Shifts.Models
         public LocalDateTime StartTime { get; set; }
         public LocalDateTime Bedtime { get; set; }
         public LocalDateTime? EndTime { get; set; }
-        public int HourlyRate { get; set; }
-        public int HourlyRateBetweenBedtimeAndMidnight { get; set; }
-        public int HourlyRateAfterMidnight { get; set; }
+        public long HourlyRate { get; set; }
+        public long HourlyRateBetweenBedtimeAndMidnight { get; set; }
+        public long HourlyRateAfterMidnight { get; set; }
     }
 }
